Move aperture photometry to a new module
