+++ photutils/aperture/__init__.py
@@ -8,4 +8,5 @@ from .circle import *  # noqa
 from .core import *  # noqa
 from .ellipse import *  # noqa
 from .mask import *  # noqa
+from .photometry import *  # noqa
 from .rectangle import *  # noqa
