+++ photutils/aperture/attributes.py
@@ -1,6 +1,7 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 """
-Descriptor classes for aperture attribute validation.
+This module defines descriptor classes for aperture attribute
+validation.
 """
 
 import warnings
