+++ photutils/aperture/tests/test_circle.py
@@ -1,4 +1,7 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
+"""
+Tests for the circle module.
+"""
 
 from astropy.coordinates import SkyCoord
 import astropy.units as u
